Add z!rw, remove z!nln, fix bug in z!e
I forgot to actually append the last line of emotes to the messages variable within z!e, so it was losing about 60 emotes. whoops. I also kept z!nln as a function nln(n), so I can z!eval it if needed, but no one else ever needed the command.