Add z!sconfig
this is the beginning stage of a command whose uses I hope to expand greatly in the near future. the first thing I've added is the ability to turn on notifications for users joining, leaving, and being banned/unbanned. I plan on adding a lot more to this, but I wanted to push out this functionality in particular as soon as I could