Fix error thrown by incoming DMs
in order to find the local command prefix, which can be customized, it checked for the guild of every message; but DMs have no guild, so they just threw an AttributeError rather than being able to run commands in DMs, like it used to. it can now run commands in DMs again.