Updating to discord.py 2.0 (and Python 3.9)
Also, finally pushing a bunch of z!pokemon updates I've had on the backburner for a while, incidentally. This will be the first of MANY updates, most likely, as I gradually go back and clean up old code, as well as update it to take advantage of some of 3.9's new features.